requirements.txt
fastapi
jinja2
pandas
geopandas
rasterio
numpy
scikit-learn
tensorflow
psycopg2-binary
python-dotenv
uvicorn
gunicorn
dvc
earthaccess
xarray
cfgrib
neuralhydrology
rioxarray
